+++ src/main/java/com/hanghae/prevstudy/domain/board/service/BoardService.java
@@ -3,7 +3,7 @@ package com.hanghae.prevstudy.domain.board.service;
 import com.hanghae.prevstudy.domain.board.dto.BoardAddRequest;
 import com.hanghae.prevstudy.domain.board.dto.BoardResponse;
 import com.hanghae.prevstudy.domain.board.dto.BoardUpdateRequest;
-import com.hanghae.prevstudy.domain.security.UserDetailsImpl;
+import com.hanghae.prevstudy.domain.security.dto.UserDetailsImpl;
 
 import java.util.List;
