+++ src/main/java/com/hanghae/prevstudy/domain/member/dto/AuthResultDto.java
@@ -1,6 +1,6 @@
 package com.hanghae.prevstudy.domain.member.dto;
 
-import com.hanghae.prevstudy.domain.security.TokenDto;
+import com.hanghae.prevstudy.domain.security.dto.TokenDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
