+++ src/main/java/com/hanghae/prevstudy/domain/member/service/MemberServiceImpl.java
@@ -1,14 +1,14 @@
 package com.hanghae.prevstudy.domain.member.service;
 
 import com.hanghae.prevstudy.domain.member.entity.Member;
-import com.hanghae.prevstudy.domain.member.exception.MemberErrorCode;
+import com.hanghae.prevstudy.global.exception.errorCode.MemberErrorCode;
 import com.hanghae.prevstudy.domain.member.repository.MemberRepository;
 import com.hanghae.prevstudy.domain.member.dto.AuthResultDto;
 import com.hanghae.prevstudy.domain.member.dto.LoginRequest;
 import com.hanghae.prevstudy.domain.member.dto.LoginResponse;
 import com.hanghae.prevstudy.domain.member.dto.SignupRequest;
-import com.hanghae.prevstudy.domain.security.TokenDto;
-import com.hanghae.prevstudy.domain.security.TokenProvider;
+import com.hanghae.prevstudy.domain.security.dto.TokenDto;
+import com.hanghae.prevstudy.domain.security.service.TokenProvider;
 import com.hanghae.prevstudy.global.exception.PrevStudyException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
