+++ src/main/java/com/hanghae/prevstudy/domain/security/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.hanghae.prevstudy.domain.security;
+package com.hanghae.prevstudy.domain.security.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
