+++ src/main/java/com/hanghae/prevstudy/domain/security/service/UserDetailsServiceImpl.java
@@ -1,7 +1,8 @@
-package com.hanghae.prevstudy.domain.security;
+package com.hanghae.prevstudy.domain.security.service;
 
 import com.hanghae.prevstudy.domain.member.entity.Member;
 import com.hanghae.prevstudy.domain.member.repository.MemberRepository;
+import com.hanghae.prevstudy.domain.security.dto.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
