+++ src/main/java/com/hanghae/prevstudy/global/config/WebMvcConfig.java
@@ -1,6 +1,6 @@
-package com.hanghae.prevstudy.domain.security;
+package com.hanghae.prevstudy.global.config;
 
-import com.hanghae.prevstudy.global.AuthMemberArgumentResolver;
+import com.hanghae.prevstudy.global.resolver.AuthMemberArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
