+++ src/test/java/com/hanghae/prevstudy/domain/member/MemberControllerTest.java
@@ -9,9 +9,9 @@ import com.hanghae.prevstudy.domain.member.dto.AuthResultDto;
 import com.hanghae.prevstudy.domain.member.dto.LoginRequest;
 import com.hanghae.prevstudy.domain.member.dto.LoginResponse;
 import com.hanghae.prevstudy.domain.member.dto.SignupRequest;
-import com.hanghae.prevstudy.domain.member.exception.MemberErrorCode;
+import com.hanghae.prevstudy.global.exception.errorCode.MemberErrorCode;
 import com.hanghae.prevstudy.domain.member.service.MemberService;
-import com.hanghae.prevstudy.domain.security.TokenDto;
+import com.hanghae.prevstudy.domain.security.dto.TokenDto;
 import com.hanghae.prevstudy.global.exception.GlobalExceptionHandler;
 import com.hanghae.prevstudy.global.exception.PrevStudyException;
 import org.hamcrest.Matchers;
