+++ src/test/java/com/hanghae/prevstudy/domain/member/MemberServiceTest.java
@@ -5,11 +5,11 @@ import com.hanghae.prevstudy.domain.member.dto.LoginRequest;
 import com.hanghae.prevstudy.domain.member.dto.LoginResponse;
 import com.hanghae.prevstudy.domain.member.dto.SignupRequest;
 import com.hanghae.prevstudy.domain.member.entity.Member;
-import com.hanghae.prevstudy.domain.member.exception.MemberErrorCode;
+import com.hanghae.prevstudy.global.exception.errorCode.MemberErrorCode;
 import com.hanghae.prevstudy.domain.member.repository.MemberRepository;
 import com.hanghae.prevstudy.domain.member.service.MemberServiceImpl;
-import com.hanghae.prevstudy.domain.security.TokenDto;
-import com.hanghae.prevstudy.domain.security.TokenProvider;
+import com.hanghae.prevstudy.domain.security.dto.TokenDto;
+import com.hanghae.prevstudy.domain.security.service.TokenProvider;
 import com.hanghae.prevstudy.global.exception.PrevStudyException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
