+++ src/test/java/com/hanghae/prevstudy/domain/security/AuthenticationTest.java
@@ -1,5 +1,8 @@
 package com.hanghae.prevstudy.domain.security;
 
+import com.hanghae.prevstudy.domain.security.dto.UserDetailsImpl;
+import com.hanghae.prevstudy.domain.security.service.TokenProvider;
+import com.hanghae.prevstudy.domain.security.service.UserDetailsServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.BDDMockito;
