+++ src/test/java/com/hanghae/prevstudy/domain/security/TokenProviderTest.java
@@ -1,5 +1,8 @@
 package com.hanghae.prevstudy.domain.security;
 
+import com.hanghae.prevstudy.domain.security.dto.TokenDto;
+import com.hanghae.prevstudy.domain.security.service.TokenProvider;
+import com.hanghae.prevstudy.global.exception.JwtValidationException;
 import io.jsonwebtoken.Claims;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
